backend/requirements.txt
fastapi
mangum
uvicorn
sqlalchemy[asyncio]
asyncpg
python-dotenv
boto3
openai